refactor: 에셋으로 변경
