+++ src/pages/SignUp/SignUpSelect/index.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import BottomArrow from '../../../assets/BottomArrow';
-import { IFormStructure, errorCheck } from '../SignUpForm/FormStructureInfo';
+import { IFormStructure, errorCheck } from '../../../types/SignUpFormType';
 import {
   SelectBoxWrapper,
   SignUpSelectBoxInner,
